% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_petersburg_game.R
\name{st_petersburg_game}
\alias{st_petersburg_game}
\title{St-Petersburg paradox games}
\usage{
st_petersburg_game(n_games, fee, seed)
}
\arguments{
\item{n_games}{number of games to play}

\item{fee}{fee for playing one game}

\item{seed}{seed for the random generator (for reproducibility)}
}
\value{
a vector of gains
}
\description{
The function returns a veector of net gains (money gained minus fee).
}
\examples{
st_petersburg_game(10,10,1)
}
\author{
Samuel Orso
}
